+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # sapper changelog
 
+## FORK
+
+* Search "FORK" with VSCode to find FORK changes
+* Change directories resolving, basing in build dir instead of cwd to allow multiple sapper instances in the same process
+* Change process.env.PORT to req.socket.localPort in server route preload to allow mutiple sapper instances in the same process but listening at different ports
+
 ## Unreleased
 
 * Fix dynamic imports in Rollup entry chunk ([#1595](https://github.com/sveltejs/sapper/issues/1595))